Adds test and util func
